+++ tests/SimpleThings/Tests/EntityAudit/CoreTest.php
@@ -35,6 +35,7 @@ class CoreTest extends BaseTest
     protected $schemaEntities = array(
         'SimpleThings\EntityAudit\Tests\Fixtures\Core\ArticleAudit',
         'SimpleThings\EntityAudit\Tests\Fixtures\Core\UserAudit',
+        'SimpleThings\EntityAudit\Tests\Fixtures\Core\ProfileAudit',
         'SimpleThings\EntityAudit\Tests\Fixtures\Core\AnimalAudit',
         'SimpleThings\EntityAudit\Tests\Fixtures\Core\Fox',
         'SimpleThings\EntityAudit\Tests\Fixtures\Core\Rabbit',
@@ -46,6 +47,7 @@ class CoreTest extends BaseTest
     protected $auditedEntities = array(
         'SimpleThings\EntityAudit\Tests\Fixtures\Core\ArticleAudit',
         'SimpleThings\EntityAudit\Tests\Fixtures\Core\UserAudit',
+        'SimpleThings\EntityAudit\Tests\Fixtures\Core\ProfileAudit',
         'SimpleThings\EntityAudit\Tests\Fixtures\Core\AnimalAudit',
         'SimpleThings\EntityAudit\Tests\Fixtures\Core\Rabbit',
         'SimpleThings\EntityAudit\Tests\Fixtures\Core\Fox',
@@ -222,6 +224,54 @@ class CoreTest extends BaseTest
         $this->assertInstanceOf('SimpleThings\EntityAudit\Tests\Fixtures\Core\UserAudit', $changedEntities[1]->getEntity());
     }
 
+    public function testNotVersionedRelationFind()
+    {
+        // Insert user without the manager to skip revision registering.
+        $this->em->getConnection()->insert(
+            $this->em->getClassMetadata('SimpleThings\EntityAudit\Tests\Fixtures\Core\UserAudit')->getTableName(),
+            array(
+                'id' => 1,
+                'name' => 'beberlei',
+            )
+        );
+
+        $article = new ArticleAudit(
+            "test",
+            "yadda!",
+            $this->em->getRepository('SimpleThings\EntityAudit\Tests\Fixtures\Core\UserAudit')->find(1),
+            'text'
+        );
+
+        $this->em->persist($article);
+        $this->em->flush();
+
+        $reader = $this->auditManager->createAuditReader($this->em);
+
+        $this->assertSame('beberlei', $reader->find(get_class($article), 1, 1)->getAuthor()->getName());
+    }
+
+    public function testNotVersionedReverseRelationFind()
+    {
+        $user = new UserAudit('beberlei');
+
+        $this->em->persist($user);
+        $this->em->flush();
+
+        // Insert user without the manager to skip revision registering.
+        $this->em->getConnection()->insert(
+            $this->em->getClassMetadata('SimpleThings\EntityAudit\Tests\Fixtures\Core\ProfileAudit')->getTableName(),
+            array(
+                'id' => 1,
+                'biography' => 'He is an amazing contributor!',
+                'user_id' => 1,
+            )
+        );
+
+        $reader = $this->auditManager->createAuditReader($this->em);
+
+        $this->assertSame('He is an amazing contributor!', $reader->find(get_class($user), 1, 1)->getProfile()->getBiography());
+    }
+
     public function testFindRevisions()
     {
         $user = new UserAudit("beberlei");
@@ -341,14 +391,14 @@ class CoreTest extends BaseTest
         $this->auditManager->getConfiguration()->setUsernameCallable(function () {
             return 'user: ' . uniqid();
         });
-        
+
         $user = new UserAudit('beberlei');
         $this->em->persist($user);
         $this->em->flush();
-        
+
         $user->setName('b.eberlei');
         $this->em->flush();
-        
+
         $reader = $this->auditManager->createAuditReader($this->em);
         $revisions = $reader->findRevisionHistory();